I am checking in with a bug fixed that I thought disabled the Login component.
The cuased turned out to be I added another component that was invisible at the time, but the space was covering the main area so the Login control was not clickable.

Set z-index to 100 brought the component to the front. Might be a little high of a number, but in front of the Welcome message is the real need.